+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 - Added a JavaScript parser generator backend, selected with --backend javascript. The generated ECMAScript module imports the token constants from the generated scanner, whose module specifier can be set with --backend-javascript-scanner-module.
 - Added a TypeScript scanner generator backend.
 - Added a TypeScript parser generator backend.
+- Added a C# scanner generator backend.
+- Added a C# parser generator backend.
 
 ## v0.2.0 (2026-08-08)